requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1
gunicorn==20.1.0
psycopg2-binary
